% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail_fd.R
\name{avail_fd}
\alias{avail_fd}
\title{Get vector of first derivatives for probability of availability}
\usage{
avail_fd(
  species = NULL,
  model = NULL,
  od = NULL,
  tssr = NULL,
  time = NULL,
  pairwise = FALSE
)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model}{Numeric or vector of model numbers ranging from 1 - 9. Can also use string "best" for best model chosen by AIC.}

\item{od}{Ordinal day, numeric digit or vector}

\item{tssr}{Time since sunrise, numeric digit or vector}

\item{time}{Maximum survey time in minutes, numeric digit or vector}

\item{pairwise}{If FALSE (default), returns a cue rate for every combination of OD and TSSR supplied;
if TRUE, returns probability of availability for each OD/TSSR pair (and so length(od) must equal length(tssr))}
}
\value{
Matrix of first derivatives with respect to each parameter, evaluated at MLE
}
\description{
\code{avail_fd} returns the first derivative of the probability of availability function,
  with respect to each parameter,
  evaluated at the maximum likelihood estimates for the given species x model combination.
  This is useful for propagating the variance in the detectability function through to
  a model of counts, if detectability is to be modelled separately from the counts.
}
\examples{
\dontrun{
# Get the probability of availability for American Robin ("AMRO"), using the best model
#   on June 1 (OD = 153), 1 hour after sunrise, for a survey of 5 minutes.
avail(species = "AMRO",
      model = "best",
      od = 153,
      tssr = 1,
      time = 5)

# Same as previous example, but this time with uncertainty, for model 7
avail(species = "AMRO",
      model = 7,
      od = 153,
      tssr = 1,
      time = 5,
      quantiles = c(0.025, 0.975))

# Availability for multiple species, multiple days, multiple models
avail(species = c("AMRO", "AMGO", "BCCH", "SCTA"),
      model = c(1,4,6,7,8),
      od = seq(90, 180, by = 2),
      tssr = 1,
      time = 5,
      quantiles = c(0.025, 0.975))

 }

}
