% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-sql-query.R
\name{build_sql_query}
\alias{build_sql_query}
\title{Build the SQL search query}
\usage{
build_sql_query(base = NULL, species = NULL, model = NULL)
}
\arguments{
\item{base}{Base search string}

\item{species}{Species to add to search string}

\item{model}{Model to add to search string}
}
\value{
String of SQL query
}
\description{
\code{build_sql_query} is an internal function that builds the SQL query to search
  the database for most tables
}
