% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data-exists.R
\name{check_data_exists}
\alias{check_data_exists}
\title{Check for existance of NA-POPS data}
\usage{
check_data_exists()
}
\value{
None
}
\description{
\code{check_data_exists} is an internal function that checks for the existance of
  NA-POPS data on the user's machine
}
