% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-for-updates.R
\name{check_for_updates}
\alias{check_for_updates}
\title{Check for updated NA-POPS results}
\usage{
check_for_updates()
}
\value{
None
}
\description{
\code{check_for_updates} checks the NA-POPS Github repository to see if there
  are updated NA-POPS results compared to what is stored locally
}
\examples{

# Check for updates to NA-POPS results
check_for_updates()

}
