% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-valid-model.R
\name{check_valid_model}
\alias{check_valid_model}
\title{Check if the models entered are valid in the NA-POPS database}
\usage{
check_valid_model(model = NULL, mod = NULL)
}
\arguments{
\item{model}{Model number}

\item{mod}{"dis" or "rem"}
}
\value{
None
}
\description{
\code{check_valid_model} is an internal function that checks for validity in the
  model arguments
}
