% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-valid-species.R
\name{check_valid_species}
\alias{check_valid_species}
\title{Check if the species exists in the NA-POPS database}
\usage{
check_valid_species(species = NULL, mod = NULL)
}
\arguments{
\item{species}{Species 4 letter code}

\item{mod}{"dis" or "rem"}
}
\value{
None
}
\description{
\code{check_valid_species} is an internal function that checks for the existence of
  the species in the NA-POPS database, given the removal or distance modelling
}
