% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-distance.R
\name{coef_distance}
\alias{coef_distance}
\title{Get coefficients of a species for a distance model}
\usage{
coef_distance(species = NULL, model = NULL)
}
\arguments{
\item{species}{4-letter banding code (or vector of) for the desired species}

\item{model}{Numeric or vector of model numbers}
}
\value{
Dataframe of coefficients for all species and distance models selected
}
\description{
\code{coef_distance} returns a dataframe of coefficients for the selected species/
  distance model combinations.
}
\examples{
\dontrun{
# Get coefficients for distance model 1 for American Robin (AMRO)
coefs <- coef_distance(species = "AMRO", model = 1)

# Get coefficients for all distance models for American Robin (AMRO)
# Option 1
coefs <- coef_distance(species = "AMRO", model = seq(1,5))
# Option 2
coefs <- coef_distance(species = "AMRO")

# Get coefficients for American Robin, Black-throated Blue Warbler, and
#  Yellow-throated Warbler for distance models 1, 4, and 5
coefs <- coef_distance(species = c("AMRO", "BTBW", "YTWA"),
                  model = c(1,4,5))

# Get coefficients for all species for distance model 5
coefs <- coef_distance(model = 5)

# Get coefficients for all species, for all distance models
coefs <- coef_distance()
}

}
