% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cue-rate.R
\name{cue_rate}
\alias{cue_rate}
\title{Get cue rate for a species}
\usage{
cue_rate(species = NULL, model = "best", od = NULL, tssr = NULL)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model}{Which model to use? Defaults to the "best" model,
but will accept the string "best" (for the best model as chosen by AIC),
"full" (for the full model with all covariates), or any numeric digit
in the range (1,9) corresponding to models 1 - 9.}

\item{od}{Ordinal day, numeric digit in the range (1,365)}

\item{tssr}{Time since sunrise, numeric digit in the range (-10,10)}
}
\value{
Numeric cue rate for species
}
\description{
\code{cue_rate} calculates the cue rate for the supplied species, given the desired model,
ordinal day, and time since sunrise.
}
\examples{

# Get the cue rate for American Robin ("AMRO"), using the best model
#   on June 1 (OD = 153), 1 hour after sunrise, for a survey of 5 minutes.
cue_rate(species = "AMRO", model = "best", od = 153, tssr = 1)

# Same as previous example, but this time with the full model
cue_rate(species = "AMRO", model = "full", od = 153, tssr = 1)

# Same as previous, but this time choosing ONLY the model that considers TSSR
cue_rate(species = "AMRO", model = 2, tssr = 1)

}
