% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cue-rate-bootstrap.R
\name{cue_rate_bootstrap}
\alias{cue_rate_bootstrap}
\title{Get cue rate for a species}
\usage{
cue_rate_bootstrap(
  vcv = NULL,
  coefficients = NULL,
  design = NULL,
  quantiles = NULL,
  samples = 1000
)
}
\arguments{
\item{quantiles}{Optional range of quantiles to calculate bootstrapped uncertainty about the estimate. Defaults to NULL}

\item{samples}{Number of bootstrap samples if bootstrapped uncertainty is to be calculated. Defaults to 1000}

\item{species}{4-letter banding code for the desired species}

\item{model}{Numeric or vector of model numbers ranging from 1 - 9.}

\item{od}{Ordinal day, numeric digit in the range (1,365)}

\item{tssr}{Time since sunrise, numeric digit in the range (-10,10)}

\item{pairwise}{If FALSE (default), returns a cue rate for every combination of OD and TSSR supplied;
if TRUE, returns cue rate for each OD/TSSR pair (and so length(od) must equal length(tssr))}
}
\value{
Numeric cue rate for species
}
\description{
\code{cue_rate} calculates the cue rate for the supplied species, given the desired model,
ordinal day, and time since sunrise.
}
\examples{

}
