% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr.R
\name{edr}
\alias{edr}
\title{Get effective detection radius for a species}
\usage{
edr(species = NULL, model = "best", roadside = NULL, forest = NULL)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model}{Which model to use? Defaults to the "best" model,
but will accept the string "best" (for the best model as chosen by AIC),
"full" (for the full model with all covariates), or any numeric digit
in the range (1,5) corresponding to models 1 - 5.}

\item{roadside}{Survey roadside status, boolean TRUE or FALSE}

\item{forest}{Forest coverage, proportion between 0 and 1}
}
\value{
Numeric effective detection radius for species
}
\description{
\code{edr} calculates the effective detection radius (EDR) for the supplied species, given the desired model,
roadside status, and forest coverage.
}
\examples{

# Get the effective detection radius for American Robin ("AMRO"), using the best model
#   for a roadside survey under 50\% forest coverage.
edr(species = "AMRO", model = "best", roadside = TRUE, forest = 0.5)

# Same as previous example, but this time with the full model
edr(species = "AMRO", model = "full", roadside = TRUE, forest = 0.5)

# Use only roadside model, and get EDR for offroad surveys
edr(species = "AMRO", model = 2, roadside = FALSE)

}
