% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-distance-coefs.R
\name{get_distance_coefs}
\alias{get_distance_coefs}
\title{Get coefficients of a species for a distance model}
\usage{
get_distance_coefs(species = NULL, model = NULL)
}
\arguments{
\item{species}{4-letter banding code (or vector of) for the desired species}

\item{model}{Numeric or vector of model numbers}
}
\value{
Dataframe of coefficients for all species and distance models selected
}
\description{
\code{get_distance_coefs} returns a dataframe of coefficients for the selected species/
  distance model combinations.
}
\examples{
# Get coefficients for distance model 1 for American Robin (AMRO)
coefs <- get_distance_coefs(species = "AMRO", model = 1)

# Get coefficients for all distance models for American Robin (AMRO)
# Option 1
coefs <- get_distance_coefs(species = "AMRO", model = seq(1,9))
# Option 2
coefs <- get_distance_coefs(species = "AMRO")

# Get coefficients for American Robin, Black-throated Blue Warbler, and
#  Yellow-throated Warbler for distance models 1, 4, and 5
coefs <- get_distance_coefs(species = c("AMRO", "BTBW", "YTWA"),
                  model = c(1,4,5))

# Get coefficients for all species for distance model 5
coefs <- get_distance_coefs(model = 5)

# Get coefficients for all species, for all distance models
coefs <- get_distance_coefs()

}
