% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-removal-coefs.R
\name{get_removal_coefs}
\alias{get_removal_coefs}
\title{Get coefficients of a species for a removal model}
\usage{
get_removal_coefs(species = NULL, model = NULL)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model}{Numeric or vector of model numbers}
}
\value{
Dataframe of coefficients for all species and removal models selected
}
\description{
\code{get_removal_coefs} returns a dataframe of coefficients for the selected species/
  removal model combinations.
}
\examples{

# Get coefficients for removal model 1 for American Robin (AMRO)
coefs <- get_removal_coefs(species = "AMRO", model = 1)

# Get coefficients for all removal models for American Robin (AMRO)
# Option 1
coefs <- get_removal_coefs(species = "AMRO", model = seq(1,9))
# Option 2
coefs <- get_removal_coefs(species = "AMRO")

# Get coefficients for American Robin, Black-throated Blue Warbler, and
#  Yellow-throated Warbler for removal models 1, 4, and 7
coefs <- get_removal_coefs(species = c("AMRO", "BTBW", "YTWA"),
                  model = c(1,4,7))

# Get coefficients for all species for removal model 9
coefs <- get_removal_coefs(model = 9)

# Get coefficients for all species, for all removal models
coefs <- get_removal_coefs()

}
