% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-removal-covariates.R
\name{get_removal_covariates}
\alias{get_removal_covariates}
\title{Get covariate space for removal modelling}
\usage{
get_removal_covariates(project = TRUE, species = NULL)
}
\arguments{
\item{project}{If TRUE, will return the set of covariates that were associated
with each sampling event in the project; otherwise, will return the covariate matrix
used for each selected species}

\item{species}{4-letter banding code (or vector of) for the desired species.
Not needed/ignored if project = TRUE}
}
\value{
Dataframe of covariates with the following entries:
  \item{OD}{Ordinal Day}
  \item{TSSR}{Time Since Sunrise}
}
\description{
\code{get_removal_covariates} returns a dataframe of covariates that were associated
  with each sampling event in NA-POPS.
}
\examples{

rem_covars <- get_removal_covariates()

}
