% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-vcv.R
\name{get_vcv}
\alias{get_vcv}
\title{Get variance-covariance matrix for a species}
\usage{
get_vcv(species = NULL, model_type = NULL, model_num = NULL)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model_type}{"rem" for removal, "dis" for distance}

\item{model_num}{Numeric or vector of model numbers ranging from 1 - 9
for model_type = "rem", or ranging from 1-5 for model_type = "dis".}
}
\value{
Variance-covariance matrix
}
\description{
\code{get_vcv} calculates the cue rate for the supplied species, given the desired model,
ordinal day, and time since sunrise.
}
