% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-projects.R
\name{list_projects}
\alias{list_projects}
\title{View a list of projects in NA-POPS}
\usage{
list_projects()
}
\value{
Data frame with the following columns
  \item{Metaproject}{Name of the umbrella organization that hosts several projects}
  \item{Project}{Name of the project}
}
\description{
\code{list_projects} returns a data frame of all projects used for NA-POPS analysis.
}
\examples{

# Get a list of projects in NA-POPS
projects <- list_projects()

}
