% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-survey-radius.R
\name{max_survey_radius}
\alias{max_survey_radius}
\title{Get maximum radius for distance survey method}
\usage{
max_survey_radius(code = NULL)
}
\arguments{
\item{code}{Alphabetical code for survey method.}
}
\value{
Numerical value of maximum survey radius, in metres
}
\description{
\code{max_survey_radius} returns the maximum survey radius, in metres, for the specified survey code
}
\examples{


}
