% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/napops-summary.R
\name{napops_summary}
\alias{napops_summary}
\title{Get summary statistics for the NA-POPS project}
\usage{
napops_summary()
}
\value{
List with the following entries:
  \item{Date}{Date that the models were run (functions as the version)}
  \item{Observations}{Number of landbird observations}
  \item{Samples}{Total number of point counts analysed}
  \item{Samples_Removal}{Total number of point counts with removal sampling}
  \item{Samples_Distance}{Total number of point counts with distance sampling}
  \item{Projects}{Total number of projects compiled in NA-POPS database}
  \item{Species}{Total number of species that have cue rates or effective detection radius derived}
}
\description{
\code{napops_summary} returns a list of summary statistics for the entire NA-POPS project.
}
\examples{

# This function can simply be called with
summary <- napops_summary()

}
