% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percept-fd.R
\name{percept_fd}
\alias{percept_fd}
\title{Get vector of first derivatives for probability of perceptibility}
\usage{
percept_fd(
  species = NULL,
  model = NULL,
  road = NULL,
  forest = NULL,
  distance = NULL,
  pairwise = FALSE
)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model}{Numeric or vector of model numbers ranging from 1 - 5. Can also use string "best" for best model chosen by AIC.}

\item{road}{Survey roadside status, boolean TRUE or FALSE}

\item{forest}{Forest coverage, proportion between 0 and 1}

\item{distance}{Distance to bird in metres}

\item{pairwise}{If FALSE (default), returns perceptibility for every combination of Road and Forest supplied;
if TRUE, returns perceptibility for each Road/Forest pair (and so length(road) must equal length(forest))}
}
\value{
Matrix of first derivatives with respect to each parameter, evaluated at MLE
}
\description{
\code{percept_fd} returns the first derivative of the probability of perceptibility function,
  with respect to each parameter,
  evaluated at the maximum likelihood estimates for the given species x model combination.
  This is useful for propagating the variance in the detectability function through to
  a model of counts, if detectability is to be modelled separately from the counts.
}
\examples{

\dontrun{
# Get the perceptibility for American Robin ("AMRO"), using the best model
#   for a roadside survey with 100\% forest coverage, bird is 100m away
percept(species = "AMRO",
        model = "best",
        road = TRUE,
        forest = 1.0,
        distance = 100)

# Same as previous example, but this time with uncertainty, for model 4
percept(species = "AMRO",
        model = 4,
        road = TRUE,
        forest = 1.0,
        distance = 100)
        quantiles = c(0.025, 0.975))

# Effective detection radius for multiple species, multiple forest coverage, multiple models
percept(species = c("AMRO", "AMGO", "BCCH", "SCTA"),
        model = c(1,4,5),
        road = TRUE,
        forest = seq(0, 1, by = 0.1),
        distance = 100)
        quantiles = c(0.025, 0.975))
}
}
