% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-coverage.R
\name{spatial_coverage}
\alias{spatial_coverage}
\title{Get spatial coverage of NA-POPS data}
\usage{
spatial_coverage(model = "all", species = NULL)
}
\arguments{
\item{model}{"rem" for removal, "dis" for distance, "all" for overall coverage (default);
cannot be a vector of models}

\item{species}{4-letter banding code (or vector of) for the desired species.
Not needed/ignored if model = "all"}
}
\value{
Simple Features (sf) object for select model x species combination
}
\description{
\code{get_spatial_coverage} returns a Simple Features(sf) object for the selected
  model x species combination. This function can also return the overall project
  coverage for all species
}
\examples{
\dontrun{
# Get the spatial coverage for the entire NA-POPS project
napops_spatial <- spatial_coverage()

# Get the spatial coverage for American Robin distance sampling
amro_dis_spatial <- spatial_coverage(model = "dis", species = "AMRO")

#' # Get the spatial coverage for Scarlet Tanager removal sampling
scta_rem_spatial <- spatial_coverage(model = "rem", species = "SCTA")
}
}
