% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-info.R
\name{survey_info}
\alias{survey_info}
\title{Get attributes for survey method}
\usage{
survey_info(code = NULL, model = NULL)
}
\arguments{
\item{code}{Alphabetical code for survey method.}

\item{model}{"rem" for removal model, "dis" for distance model}
}
\value{
List with the following entries:
  \item{Method}{Alphabetical code for the survey code}
  \item{Model}{"rem" or "dis" depending on user's entry}
  \item{Description}{Description of the distance or time bins in this method}
  \item{Max_Interval}{Maximum time or distance considered for this method}
  \item{Units}{"Minutes" if removal model selected, "Metres" if distance model selected}
}
\description{
\code{survey_info} returns a list of attributes associated with the survey method code.
}
\examples{


}
