% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail.R
\name{avail}
\alias{avail}
\title{Get probability of availability for species}
\usage{
avail(species = NULL, model = "best", od = NULL, tssr = NULL, time = NULL)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model}{Which model to use? Defaults to the "best" model,
but will accept the string "best" (for the best model as chosen by AIC),
"full" (for the full model with all covariates), or any numeric digit
in the range (1,9) corresponding to models 1 - 9.}

\item{od}{Ordinal day, numeric digit in the range (1,365)}

\item{tssr}{Time since sunrise, numeric digit in the range (-10,10)}

\item{time}{Maximum survey time in minutes}
}
\value{
Probability of availability
}
\description{
\code{avail} calculates the probability that the specified species will give
  a cue within a maximum time range, given the day of the year and time of day.
}
