% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avail.R
\name{avail}
\alias{avail}
\title{Get probability of availability for species}
\usage{
avail(
  species = NULL,
  model = NULL,
  od = NULL,
  tssr = NULL,
  time = NULL,
  pairwise = FALSE,
  quantiles = NULL,
  samples = 1000
)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model}{Numeric or vector of model numbers ranging from 1 - 9.}

\item{od}{Ordinal day, numeric digit or vector}

\item{tssr}{Time since sunrise, numeric digit or vector}

\item{time}{Maximum survey time in minutes, numeric digit or vector}

\item{pairwise}{If FALSE (default), returns a cue rate for every combination of OD and TSSR supplied;
if TRUE, returns probability of availability for each OD/TSSR pair (and so length(od) must equal length(tssr))}

\item{quantiles}{Optional range of quantiles to calculate bootstrapped uncertainty about the estimate. Defaults to NULL}

\item{samples}{Number of bootstrap samples if bootstrapped uncertainty is to be calculated. Defaults to 1000}
}
\value{
Probability of availability
}
\description{
\code{avail} calculates the probability that the specified species will give
  a cue within a maximum time range, given the day of the year and time of day.
}
\examples{

}
