% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap estimates, internal function}
\usage{
bootstrap(
  vcv = NULL,
  coefficients = NULL,
  design = NULL,
  quantiles = NULL,
  samples = 1000,
  model = NULL
)
}
\arguments{
\item{vcv}{Variance-covariance matrix}

\item{coefficients}{Point estimates of coefficients to bootstrap around}

\item{design}{Design matrix to produce values}

\item{quantiles}{Vector of quantiles to return bootstrapped point estimates of}

\item{samples}{Number of bootstrap samples. Defaults to 1000}

\item{model}{Which modelling technique, "rem" or "dis"}
}
\value{
Numeric cue rate for species
}
\description{
\code{bootstrap} calculates a bootstrapped distribution for either cue rate or effective
  detection radius. Internal function.
}
\examples{

}
