% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cue-rate.R
\name{cue_rate}
\alias{cue_rate}
\title{Get cue rate for a species}
\usage{
cue_rate(
  species = NULL,
  model = NULL,
  od = NULL,
  tssr = NULL,
  pairwise = FALSE,
  quantiles = NULL,
  samples = 1000
)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model}{Numeric or vector of model numbers ranging from 1 - 9. Can also use string "best" for best model chosen by AIC.}

\item{od}{Ordinal day, numeric digit or vector}

\item{tssr}{Time since sunrise, numeric digit or vector}

\item{pairwise}{If FALSE (default), returns a cue rate for every combination of OD and TSSR supplied;
if TRUE, returns cue rate for each OD/TSSR pair (and so length(od) must equal length(tssr))}

\item{quantiles}{Optional range of quantiles to calculate bootstrapped uncertainty about the estimate. Defaults to NULL}

\item{samples}{Number of bootstrap samples if bootstrapped uncertainty is to be calculated. Defaults to 1000}
}
\value{
Numeric cue rate for species
}
\description{
\code{cue_rate} calculates the cue rate for the supplied species, given the desired model,
ordinal day, and time since sunrise.
}
\examples{
\dontrun{
# Get the cue rate for American Robin ("AMRO"), using the best model
#   on June 1 (OD = 153), 1 hour after sunrise.
cue_rate(species = "AMRO",
         model = "best",
         od = 153,
         tssr = 1)

# Same as previous example, but this time with uncertainty, for model 7
cue_rate(species = "AMRO",
         model = 7,
         od = 153,
         tssr = 1,
         quantiles = c(0.025, 0.975))

# Cue rate for multiple species, multiple days, multiple models
cue_rate(species = c("AMRO", "AMGO", "BCCH", "SCTA"),
         model = c(1,4,6,7,8),
         od = seq(90, 180, by = 2),
         tssr = 1,
         quantiles = c(0.025, 0.975))
}
}
