% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr.R
\name{edr}
\alias{edr}
\title{Get effective detection radius for a species}
\usage{
edr(species = NULL, model = "best", roadside = NULL, forest = NULL)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model}{Which model to use? Defaults to the "best" model,
but will accept the string "best" (for the best model as chosen by AIC),
"full" (for the full model with all covariates), or any numeric digit
in the range (1,5) corresponding to models 1 - 5.}

\item{roadside}{Survey roadside status, boolean TRUE or FALSE}

\item{forest}{Forest coverage, proportion between 0 and 1}
}
\value{
Numeric effective detection radius for species
}
\description{
\code{cue_rate} calculates the effective detection radius (EDR) for the supplied species, given the desired model,
roadside status, and forest coverage.
}
