% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edr.R
\name{edr}
\alias{edr}
\title{Get effective detection radius for a species}
\usage{
edr(
  species = NULL,
  model = NULL,
  road = NULL,
  forest = NULL,
  pairwise = FALSE,
  quantiles = NULL,
  samples = 1000
)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model}{Numeric or vector of model numbers ranging from 1 - 5. Can also use string "best" for best model chosen by AIC.}

\item{road}{Survey roadside status, boolean TRUE or FALSE}

\item{forest}{Forest coverage, proportion between 0 and 1}

\item{pairwise}{If FALSE (default), returns an effective detection radius for
every combination of Road and Forest supplied; if TRUE, returns an effective detection
radius for each Road/Forest pair (and so length(road) must equal length(forest))}

\item{quantiles}{Optional range of quantiles to calculate bootstrapped uncertainty about the estimate. Defaults to NULL}

\item{samples}{Number of bootstrap samples if bootstrapped uncertainty is to be calculated. Defaults to 1000}
}
\value{
Numeric effective detection radius for species
}
\description{
\code{edr} calculates the effective detection radius (EDR) for the supplied species, given the desired model,
roadside status, and forest coverage.
}
\examples{
\dontrun{
# Get the effective detection radius for American Robin ("AMRO"), using the best model
#   for a roadside survey with 100\% forest coverage
edr(species = "AMRO",
    model = "best",
    road = TRUE,
    forest = 1.0)

# Same as previous example, but this time with uncertainty, for model 4
edr(species = "AMRO",
    model = 4,
    road = TRUE,
    forest = 1.0
    quantiles = c(0.025, 0.975))

# Effective detection radius for multiple species, multiple forest coverage, multiple models
edr(species = c("AMRO", "AMGO", "BCCH", "SCTA"),
    model = c(1,4,5),
    road = TRUE,
    forest = seq(0, 1, by = 0.1)
    quantiles = c(0.025, 0.975))
}
}
