% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch NA-POPS modelling results}
\usage{
fetch_data(quiet = TRUE)
}
\arguments{
\item{quiet}{Logical: should download progress be suppressed? Defaults to TRUE}
}
\value{
None
}
\description{
\code{fetch_data} downloads all coefficients and covariance matrices
generated by NA-POPS models. A package-specific directory is created on the user's
computer (see documentation of \code{rappdirs::appdir} for details of where this
directory lives), and NA-POPS data is saved to that directory for use by other functions.
}
