% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percept.R
\name{percept}
\alias{percept}
\title{Get conditional probability of perceiving a bird, given availability}
\usage{
percept(
  species = NULL,
  model = "best",
  roadside = NULL,
  forest = NULL,
  distance = NULL
)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model}{Which model to use? Defaults to the "best" model,
but will accept the string "best" (for the best model as chosen by AIC),
"full" (for the full model with all covariates), or any numeric digit
in the range (1,5) corresponding to models 1 - 5.}

\item{roadside}{Survey roadside status, boolean TRUE or FALSE}

\item{forest}{Forest coverage, proportion between 0 and 1}

\item{distance}{Maximum survey distance in metres}
}
\value{
Probability of perceptibility
}
\description{
\code{percept} calculates the conditional probability that a bird is perceived
  by an observer, given that it gives a cue, dependent on the roadside status
  of a survey, forest coverage, and maximum survey distance.
}
\examples{

# Get the probability of perceptibility for American Robin ("AMRO"), using the best model
#   for a roadside survey under 50\% forest coverage with maximum survey distance of 200m.
percept(species = "AMRO", model = "best", roadside = TRUE, forest = 0.5, distance = 200)

# Same as previous example, but this time with the full model
percept(species = "AMRO", model = "full", roadside = TRUE, forest = 0.5, distance = 200)

# Use only roadside model, and specify offroad surveys
percept(species = "AMRO", model = 2, roadside = FALSE, distance = 200)

}
