% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percept.R
\name{percept}
\alias{percept}
\title{Get conditional probability of perceiving a bird, given availability}
\usage{
percept(
  species = NULL,
  model = NULL,
  road = NULL,
  forest = NULL,
  distance = NULL,
  pairwise = FALSE,
  quantiles = NULL,
  samples = 1000
)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model}{Numeric or vector of model numbers ranging from 1 - 9.}

\item{road}{Survey roadside status, boolean TRUE or FALSE}

\item{forest}{Forest coverage, proportion between 0 and 1}

\item{distance}{Maximum survey distance in metres}

\item{pairwise}{If FALSE (default), returns perceptibility for every combination of Road and Forest supplied;
if TRUE, returns perceptibility for each Road/Forest pair (and so length(road) must equal length(forest))}

\item{quantiles}{Optional range of quantiles to calculate bootstrapped uncertainty about the estimate. Defaults to NULL}

\item{samples}{Number of bootstrap samples if bootstrapped uncertainty is to be calculated. Defaults to 1000}
}
\value{
Probability of perceptibility
}
\description{
\code{percept} calculates the conditional probability that a bird is perceived
  by an observer, given that it gives a cue, dependent on the roadside status
  of a survey, forest coverage, and maximum survey distance.
}
\examples{

# Get the probability of perceptibility for American Robin ("AMRO"), using Model 2
# Use only roadside model, and specify offroad surveys
percept(species = "AMRO", model = 2, road = FALSE, distance = 200)

}
