% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percept.R
\name{percept}
\alias{percept}
\title{Get conditional probability of perceiving a bird, given the bird gives a cue}
\usage{
percept(
  species = NULL,
  model = NULL,
  road = NULL,
  forest = NULL,
  distance = NULL,
  pairwise = FALSE,
  quantiles = NULL,
  samples = 1000
)
}
\arguments{
\item{species}{4-letter banding code for the desired species}

\item{model}{Numeric or vector of model numbers ranging from 1 - 5. Can also use string "best" for best model chosen by AIC.}

\item{road}{Survey roadside status, boolean TRUE or FALSE}

\item{forest}{Forest coverage, proportion between 0 and 1}

\item{distance}{Distance to bird in metres}

\item{pairwise}{If FALSE (default), returns perceptibility for every combination of Road and Forest supplied;
if TRUE, returns perceptibility for each Road/Forest pair (and so length(road) must equal length(forest))}

\item{quantiles}{Optional range of quantiles to calculate bootstrapped uncertainty about the estimate. Defaults to NULL}

\item{samples}{Number of bootstrap samples if bootstrapped uncertainty is to be calculated. Defaults to 1000}
}
\value{
Probability of perceptibility
}
\description{
\code{percept} calculates the conditional probability that a bird is perceived
  by an observer, given that it gives a cue, dependent on the roadside status
  of a survey, forest coverage, and maximum survey distance.
}
\examples{

# Get the perceptibility for American Robin ("AMRO"), using the best model
#   for a roadside survey with 100\% forest coverage, bird is 100m away
percept(species = "AMRO",
        model = "best",
        road = TRUE,
        forest = 1.0,
        distance = 100)

# Same as previous example, but this time with uncertainty, for model 4
percept(species = "AMRO",
        model = 4,
        road = TRUE,
        forest = 1.0,
        distance = 100)
        quantiles = c(0.025, 0.975))

# Effective detection radius for multiple species, multiple forest coverage, multiple models
percept(species = c("AMRO", "AMGO", "BCCH", "SCTA"),
        model = c(1,4,5),
        road = TRUE,
        forest = seq(0, 1, by = 0.1),
        distance = 100)
        quantiles = c(0.025, 0.975))

}
